+++ src/main/java/subway/line/LineController.java
@@ -3,6 +3,7 @@ package subway.line;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.station.StationService;
 
 import javax.annotation.Resource;
 import java.util.List;
@@ -11,7 +12,11 @@ import java.util.List;
 public class LineController {
 
     @Resource
+    private StationService stationService;
+    @Resource
     private LineService lineService;
+    @Resource
+    private SectionService sectionService;
 
     @PostMapping(value = "/lines")
     public ResponseEntity<LineResponse> createLine(@RequestBody LineRequest lineRequest) {
@@ -23,28 +28,29 @@ public class LineController {
         return lineService.getLines();
     }
 
-    @DeleteMapping("/lines/{id}")
-    public ResponseEntity deleteLine(@PathVariable Long id) {
-        return lineService.delete(id);
+    @DeleteMapping("/lines/{lineId}")
+    public ResponseEntity deleteLine(@PathVariable Long lineId) {
+        return lineService.delete(lineId);
     }
 
-    @GetMapping(value = "/lines/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<LineResponse> showLine(@PathVariable Long id) {
-        return lineService.getLine(id);
+    @GetMapping(value = "/lines/{lineId}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<LineResponse> showLine(@PathVariable Long lineId) {
+        return lineService.getLine(lineId);
     }
 
-    @PutMapping(value = "/lines/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity updateLine(@PathVariable Long id, @RequestBody LineRequest lineRequest) {
-        return lineService.update(id, lineRequest);
+    @PutMapping(value = "/lines/{lineId}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity updateLine(@PathVariable Long lineId, @RequestBody LineRequest lineRequest) {
+        return lineService.update(lineId, lineRequest);
     }
 
-    @PostMapping(value = "/lines/{id}/sections")
-    public ResponseEntity createSection(@PathVariable Long id, @RequestBody SectionRequest sectionRequest) {
-        return lineService.createSection(id, sectionRequest);
+    @PostMapping(value = "/lines/{lineId}/sections")
+    public ResponseEntity createSection(@PathVariable Long lineId, @RequestBody SectionRequest sectionRequest) {
+        sectionService.validate(lineId, sectionRequest, stationService.getStations(lineId));
+        return sectionService.create(lineId, sectionRequest);
     }
 
-    @DeleteMapping(value = "/lines/{id}/sections")
-    public ResponseEntity deleteSection(@PathVariable Long id, @RequestParam Long stationId) {
-        return lineService.deleteSection(id, stationId);
+    @DeleteMapping(value = "/lines/{lineId}/sections")
+    public ResponseEntity deleteSection(@PathVariable Long lindId, @RequestParam Long stationId) {
+        return sectionService.delete(lindId, stationId);
     }
 }
